Update deep_learning_ff.tex
Updated logsumexp import for NumpyLogLinear exercise

The example in the guide was not in sync with the code in the
master/student branches. The guide was importing the logsumexp
function from scipy, while the code version uses the LxMLS version
of logsumexp. The latter is the expected behavior, so this commit 
updates the guide version.